requirements.txt
tabulate==0.9.0
sympy==1.11.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
plotly==5.14.1
st-pages==0.4.1
streamlit-extras==0.2.7
streamlit-code-editor==0.1.8
